kartochka ofline/requirements.txt
requests==2.31.0
urllib3==2.2.1
openpyxl==3.1.2
pdfplumber==0.11.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
docx2txt==0.8
